[feat]: custom snnTorch LIF for moving average tracking (#23)
Changes:
1. The previous abstractions we added for `TemporalFilter`
`SpikeMovingAverage` and `VarianceMovingAverage` are all for floats. In
order to apply these efficiently for all neurons / synapses in a layer
at once, we need to operate on matrices rather than float values (making
sure to handle both batch and feature dims).
2. Custom implementation of snnTorch LIF to encapsulate the logic of
tracking moving averages for spike and variance for a LIF layer. The
point being that we don't want to have this logic inside the network
implementation, but instead it can be plumbed down into a specific LIF
impl.
3. Hinting at units from Zenke paper. I am using one "unit" == 1sec. So
600 seconds is 6 * 10^5. We should confirm we don't want to instead make
one "unit" == 1ms. @mstormbull FYA