[feat]: LPL rule integrated into skeleton network (#24)
Integrating the LPL rule into the skeleton network. This runs
successfully, but still need to assess selectivity metric.

Changes:

1. Fixed some bugs with datagen. Namely the cluster generation needs to
start randomly from any potential cluster rather than starting
automatically from 0th. Also started applying some normalization.
2. Added the temporal filters to the network. We are using one epsilon
filter and _two_ separate alpha filters.

```
        self.alpha_filter_first_term = TemporalFilter(
            tau_rise=TAU_RISE_ALPHA, tau_fall=TAU_FALL_ALPHA)
        self.epsilon_filter_second_term = TemporalFilter(
            tau_rise=TAU_RISE_EPSILON, tau_fall=TAU_FALL_EPSILON)
        self.alpha_filter_second_term = TemporalFilter(
            tau_rise=TAU_RISE_ALPHA, tau_fall=TAU_FALL_ALPHA)
```
3. Implemented the LPL rule. When implementing the LPL rule, there is a
discrepancy with the first hidden layer, which is that it relies on
spike, mem, and variance of previous layer spiking. This can be solved
by Poisson encoding the input data, as well as applying our spike and
variance filters to the incoming input data for our first hidden layer.
For example, this is in the network init, which is responsible for
tracking these intermediary values for the learning rule.
```
        self.data_spike_moving_average = SpikeMovingAverage(
            batch_size=layer_settings.batch_size, data_size=self.layer_settings.data_size)
        self.variance_moving_average = VarianceMovingAverage()
```