[feat]: logging for pointcloud benchmark (#34)
I merged another PR which fixed a bug in my forward pass implementation,
and also scaled the tau constants (same as Zenke but there is nuance in
interpreting the units seconds vs ms -- I changed to 1 unit == 1
second).
https://github.com/and-rewsmith/LPL-SNN/pull/33

After this change I made some wandb plots around how the excitatory
neuron becomes selective to the appropriate data input in the point
cloud task. To remind about the point cloud task, 2 data points are
passed in sequentially to the model. There are two clusters for each of
these data points to fall into. The clusters are aligned with the x
coordinate range for a single cluster being very narrow, whereas the y
coordinate range is unrestricted.

The two data points at each timestep represent x,y coordinates, where
between timesteps the x coordinate usually stays within the same cluster
(small change), whereas the y coordinate changes completely randomly
with no restriction . Here is a visualization of the clustering.

![image](https://github.com/and-rewsmith/LPL-SNN/assets/19913741/2e3ed65b-54a0-4ade-9390-6ec98563ed04)


There is no output to our LPL network (1 excitatory neuron / 1
inhibitory neuron), but based on Zenke experimental results we should
see the single excitatory neuron to become selective to the x coordinate
over the y. coordinate. It looks like this happens, which is promising.
I applied some smoothing to the layer_0_excitatory spike so it is clear
that the excitatory spike trends the data_point_0 (x). But wait -- this
doesn't really mean much since of course the excitatory spike will trend
the data point 0 since the excitatory spike roughly would just follow
the total current in an untrained network. So we need more analysis.

(Another cool thing to note, is that we got past the barrier that the
excitatory neuron spiking ran away. This was due to the constant
rescaling, along with the inhibitory neurons. We are not even using
ALIFs, but I don't think Zenke actually did according to the erratum.

![image](https://github.com/and-rewsmith/LPL-SNN/assets/19913741/247e47bc-dd0e-4140-9300-2fc4e880723b)



> But wait -- this doesn't really mean much since of course the
excitatory spike will trend the data point 0 since the excitatory spike
roughly would just follow the total current in an untrained network. So
we need more analysis.

In order to analyze this further, we can check to see if the weight for
the connection between the data point 0 and the excitatory neuron
remains positive, and the weight for the connection between data point 1
and the excitatory neuron decreases. Luckily, it looks like this happens
too -- and the weight for data point 1 actually decays completely!

![image](https://github.com/and-rewsmith/LPL-SNN/assets/19913741/8bb16886-18e2-4fb0-acfb-d9664681da09)



My next thinking is that we need to clamp these weights from flipping
sign.